Move handle table stuff out of gcheaputilities.
